Move to python 3, add option to modify Catalog.xml
Move script to python 3, since python 2 is end of life. Only small
modifications required for this.
 * `Queue` module renamed to `queue`.
 * `threading.isAlive()` renamed to `threading.is_alive()`.
 * replaced subprocess/md5 with `hashlib` routine instead. Not strictly
necessary for the python3 migration, but had to do with
string encoding mess from subprocess-output.

Options to change Catalog.xml was added to remove the occurance of
`baseLocation`and `baseLocationAccessProtocols` which are present in the
original Catalog from dell. These two variables forces
lifecyclecontroller / idrac to fetch updates from the location specified
there using that protocol, instead of the mirror this script created. As
described in https://github.com/littlejon/dell-update-mirror/issues/1